+++ src/clj/qbits/jet/client/http.clj
@@ -2,6 +2,7 @@
   (:refer-clojure :exclude [get])
   (:require
    [clojure.core.async :as async]
+   [clojure.core.async.impl.protocols :as protocols]
    [qbits.jet.client.ssl :as ssl]
    [qbits.jet.client.auth :as auth]
    [qbits.jet.client.cookies :as cookies]
@@ -9,38 +10,40 @@
    [cheshire.core :as json]
    [clojure.xml :as xml])
   (:import
-   (org.eclipse.jetty.client
-    HttpClient
-    HttpRequest)
-   (org.eclipse.jetty.util Fields)
-   (org.eclipse.jetty.client.util
-    StringContentProvider
-    BytesContentProvider
-    ByteBufferContentProvider
-    DeferredContentProvider
-    InputStreamContentProvider
-    PathContentProvider
-    FormContentProvider)
-   (org.eclipse.jetty.http
-    HttpFields
-    HttpField)
-   (org.eclipse.jetty.client.api
-    Request$FailureListener
-    Response$CompleteListener
-    Response$ContentListener
-    Response$HeadersListener
-    Request
-    ;; Response
-    Result)
-   (java.util.concurrent TimeUnit)
-   (java.nio ByteBuffer)
-   (java.io
-    ByteArrayInputStream
-    ByteArrayOutputStream)
-   (clojure.lang Keyword Sequential)))
+    (org.eclipse.jetty.client HttpClient)
+    (org.eclipse.jetty.util Fields)
+    (org.eclipse.jetty.client.util
+      StringContentProvider
+      BytesContentProvider
+      ByteBufferContentProvider
+      DeferredContentProvider
+      InputStreamContentProvider
+      PathContentProvider
+      FormContentProvider)
+    (org.eclipse.jetty.http
+      HttpFields
+      HttpField)
+    (org.eclipse.jetty.client.api
+      Authentication$Result
+      Request$FailureListener
+      Response$CompleteListener
+      Response$HeadersListener
+      Request
+      ;; Response
+      Response$AsyncContentListener
+      Result)
+    (java.util.concurrent TimeUnit)
+    (java.nio ByteBuffer)
+    (java.io
+      ByteArrayInputStream
+      ByteArrayOutputStream
+      IOException)
+    (clojure.lang Sequential)))
 
 (def ^:const array-class (class (clojure.core/byte-array 0)))
 (def default-buffer-size (* 1024 1024 4))
+(def default-insecure-ssl-context-factory
+  (delay (ssl/insecure-ssl-context-factory)))
 
 (defn ^:no-doc ^array-class
   byte-buffer->bytes
@@ -49,50 +52,64 @@
     (.get bb ba)
     ba))
 
-(defn ^:no-doc byte-buffer->string
-  [^ByteBuffer bb]
-  (String. ^bytes (byte-buffer->bytes bb) "UTF-8"))
+(defn ^:no-doc byte-array->string
+  [^bytes bb]
+  (String. bb "UTF-8"))
 
 (defn ^:no-doc decode-body [bb as]
   (case as
-    :bytes (byte-buffer->bytes bb)
-    :input-stream (ByteArrayInputStream. (byte-buffer->bytes bb))
-    :string (byte-buffer->string bb)
-    :json (json/parse-string (byte-buffer->string bb) true)
-    :json-str (json/parse-string (byte-buffer->string bb) false)
-    :xml (xml/parse (ByteArrayInputStream. (byte-buffer->bytes bb)))
-    bb))
+    :bytes bb
+    :input-stream (ByteArrayInputStream. bb)
+    :string (byte-array->string bb)
+    :json (json/parse-string (byte-array->string bb) true)
+    :json-str (json/parse-string (byte-array->string bb) false)
+    :xml (xml/parse (ByteArrayInputStream. bb))
+    (ByteBuffer/wrap bb)))
 
-(defn fold-chunks+decode-xform [as]
+(defn fold-chunks+decode-xform [as buffer-size]
   (fn [reduction-function]
-    (let [ba (ByteArrayOutputStream.)]
+    (let [ba (ByteArrayOutputStream.)
+          state (atom ::open)
+          retrieve-data (fn [] (decode-body (.toByteArray ba) as))]
       (fn
+        ([] (reduction-function))
         ([result]
-         (reduction-function result (decode-body (ByteBuffer/wrap (.toByteArray ba))
-                                          as)))
+         (when (compare-and-set! state ::open ::closed)
+           (reduction-function result (retrieve-data))))
         ([result chunk]
-         (.write ba (byte-buffer->bytes chunk)))))))
+         (.write ba chunk)
+         (if (>= (.size ba) buffer-size)
+           (let [data (retrieve-data)]
+             (.reset ba)
+             (reduction-function result data))
+           (reduction-function result)))))))
 
 (defn decode-chunk-xform
   [as]
   (fn [reduction-function]
     (fn
+      ([] (reduction-function))
       ([result]
        (reduction-function result))
       ([result chunk]
        (reduction-function result (decode-body chunk as))))))
 
-(defrecord Response [status headers body])
+(defrecord Response [status headers body error-chan])
 
 (defn ^:no-doc result->response
   [^Result result body-ch]
-  (let [response (.getResponse result)]
+  (let [response (.getResponse result)
+        error-ch (async/promise-chan)]
+    (if (.isSucceeded result)
+      (async/close! error-ch)
+      (async/put! error-ch {:error (.getFailure result)}))
     (Response. (.getStatus response)
                (reduce (fn [m ^HttpField h]
                          (assoc m (string/lower-case  (.getName h)) (.getValue h)))
                        {}
                        ^HttpFields (.getHeaders response))
-               body-ch)))
+               body-ch
+               error-ch)))
 
 (defprotocol PRequest
   (encode-chunk [x])
@@ -105,7 +122,7 @@
     (ByteBuffer/wrap x))
   (encode-body [ba]
     (BytesContentProvider.
-     (into-array array-class [ba])))
+      (into-array array-class [ba])))
 
   ByteBuffer
   (encode-chunk [x] x)
@@ -150,9 +167,9 @@
 
   Object
   (encode-chunk [x]
-    (throw (ex-info "Chunk type no supported by encoder")))
+    (throw (ex-info "Chunk type no supported by encoder" {})))
   (encode-body [x]
-    (throw (ex-info "Body type no supported by encoder")))
+    (throw (ex-info "Body type no supported by encoder" {})))
   (encode-content-type [content-type]
     (encode-content-type (subs (str content-type) 1))))
 
@@ -173,6 +190,7 @@
   ([{:keys [url
             address-resolution-timeout
             connect-timeout
+            executor
             follow-redirects?
             max-redirects
             idle-timeout
@@ -194,7 +212,7 @@
           follow-redirects? true
           tcp-no-delay? true
           strict-event-ordering? false
-          ssl-context-factory ssl/insecure-ssl-context-factory
+          ssl-context-factory @default-insecure-ssl-context-factory
           request-buffer-size default-buffer-size
           response-buffer-size default-buffer-size}
      :as r}]
@@ -236,6 +254,9 @@
      (when cookie-store
        (.setCookieStore client cookie-store))
 
+     (when executor
+       (.setExecutor client executor))
+
      (.setRemoveIdleDestinations client remove-idle-destinations?)
      (.setDispatchIO client dispatch-io?)
      (.setFollowRedirects client follow-redirects?)
@@ -259,16 +280,20 @@
            timeout
            agent
            follow-redirects?
-           fold-chunked-response?]
+           fold-chunked-response?
+           fold-chunked-response-buffer-size
+           ^Authentication$Result auth]
     :or {method :get
          as :string
          follow-redirects? true
-         fold-chunked-response? true}
+         fold-chunked-response? true
+         fold-chunked-response-buffer-size Integer/MAX_VALUE}
     :as request-map}]
-  (let [ch (async/chan 1)
+  (let [ch (async/promise-chan)
+        error-ch (async/promise-chan)
         body-ch (async/chan 1
                             (if fold-chunked-response?
-                              (fold-chunks+decode-xform as)
+                              (fold-chunks+decode-xform as fold-chunked-response-buffer-size)
                               (decode-chunk-xform as)))
         request ^Request (.newRequest client ^String url)]
 
@@ -305,13 +330,24 @@
       (.header request "Content-Type"
                (name content-type)))
 
-    (doseq [[k v] query-string]
-      (.param request (name k) (str v)))
+    (when auth
+      (.apply auth request))
 
-    (.onResponseContent request
-                        (reify Response$ContentListener
-                          (onContent [this response bytebuffer]
-                            (async/put! body-ch bytebuffer))))
+    (doseq [[k v] query-string]
+      (if (coll? v)
+        (doseq [v' v]
+          (.param request (name k) (str v')))
+        (.param request (name k) (str v))))
+
+    (.onResponseContentAsync request
+                             (reify Response$AsyncContentListener
+                               (onContent [this response bytebuffer callback]
+                                 (if (protocols/closed? body-ch)
+                                   (let [ex (IOException. "Body channel closed unexpectedly")]
+                                     (.failed callback ex)
+                                     (.abort request ex))
+                                   (async/put! body-ch (byte-buffer->bytes bytebuffer)
+                                               (fn [_] (.succeeded callback)))))))
 
     (.onResponseHeaders request
                         (reify Response$HeadersListener
@@ -319,18 +355,32 @@
                             (async/put! ch
                                         (Response. (.getStatus response)
                                                    (reduce (fn [m ^HttpField h]
-                                                             (assoc m (string/lower-case (.getName h)) (.getValue h)))
+                                                             (let [k (string/lower-case (.getName h))
+                                                                   v (.getValue h)]
+                                                               (if (contains? m k)
+                                                                 (if (coll? (m k))
+                                                                   (assoc m k (conj (m k) v))
+                                                                   (assoc m k [(m k) v]))
+                                                                 (assoc m k v))))
                                                            {}
                                                            ^HttpFields (.getHeaders response))
-                                                   body-ch)))))
+                                                   body-ch
+                                                   error-ch)))))
+
+    (.onRequestFailure request
+                       (reify Request$FailureListener
+                         (onFailure [_ _ throwable]
+                           (async/put! ch {:error throwable}))))
 
     (.send request
            (reify Response$CompleteListener
-             (onComplete [this result]
-               (if (not (.isSucceeded ^Result result))
-                 (async/put! ch {:error (.getFailure result)}))
+             (onComplete [_ result]
+               (when (not (.isSucceeded ^Result result))
+                   (async/put! ch {:error (.getFailure result)})
+                   (async/put! error-ch {:error (.getFailure result)}))
                (async/close! body-ch)
-               (async/close! ch))))
+               (async/close! ch)
+               (async/close! error-ch))))
     ch))
 
 (defn get