+++ src/clj/qbits/jet/websocket.clj
@@ -61,8 +61,6 @@
   ;; (-send! [this ws] ()
   )
 
-(defrecord WebSocketBinaryFrame [payload offset len])
-
 (deftype WebSocket
     [^ManyToManyChannel in
      ^ManyToManyChannel out
@@ -118,16 +116,17 @@
     (a/put! in message #(backpressure! this %)))
 
   (onWebSocketBinary [this payload offset len]
-    (a/put! in (WebSocketBinaryFrame. payload offset len)
+    (a/put! in (ByteBuffer/wrap payload offset len)
             #(backpressure! this %)))
 
   BackPressure
   (backpressure! [this backpressure?]
-    (if backpressure?
-      (set! reads-suspend-token (.suspend session))
-      (do
-        (.resume reads-suspend-token)
-        (set! reads-suspend-token nil))))
+    (when reads-suspend-token
+      (if backpressure?
+        (set! reads-suspend-token (.suspend session))
+        (do
+          (.resume reads-suspend-token)
+          (set! reads-suspend-token nil)))))
 
   PWebSocket
   (remote [this]